+++ executables/LUDUMDARE/Ludum41Custom/src/Ludum41CustomGameInstance.h
@@ -73,13 +73,13 @@ public:
 	/** returns the current combo multiplier */
 	int GetCurrentComboMultiplier() const { return combo_multiplier; }
 
-	void TickBrickOffset(double delta_time);
+	void TickBrickOffset(float delta_time);
 
-	void TickChallenge(double delta_time);
+	void TickChallenge(float delta_time);
 
-	void TickBallSplit(double delta_time);
+	void TickBallSplit(float delta_time);
 
-	void TickBackgroundFillRatio(double delta_time);
+	void TickBackgroundFillRatio(float delta_time);
 
 	virtual void FillUniformProvider(chaos::GPUProgramProvider& main_uniform_provider) override;
 
@@ -99,7 +99,7 @@ protected:
 
 	virtual death::Player * DoCreatePlayer() override;
 
-	virtual bool DoTick(double delta_time) override;
+	virtual bool DoTick(float delta_time) override;
 
 	virtual void OnLevelChanged(death::GameLevel * new_level, death::GameLevel * old_level, death::GameLevelInstance * new_level_instance) override;
